Update xcodebuild signing arguments
Updates the signing code to change 'PROVISIONING_PROFILE' to
'PROVISIONING_PROFILE_APP' and 'PROVISIONING_PROFILE_SPECIFIER' to
'PROVISIONING_PROFILE_SPECIFIER_APP'.

Reason is Unity starting from version 2019.3 is including a new framework
'UnityFramework.framework' and that does not support signing. Using
the 'APP' suffix makes sure it only applies to the application target
and not all targets.

More information can be found in the Unity manual:
https://docs.unity3d.com/Manual/StructureOfXcodeProject.html